+++ packages/cli/src/ui/App.tsx
@@ -166,7 +166,7 @@ const App = ({
   const { stats: sessionStats } = useSessionStats();
   const [staticNeedsRefresh, setStaticNeedsRefresh] = useState(false);
   const [staticKey, setStaticKey] = useState(0);
-  const refreshStatic = useCallback(() => {
+  const refreshConfig = useCallback(() => {
     stdout.write(ansiEscapes.clearTerminal);
     setStaticKey((prev) => prev + 1);
   }, [setStaticKey, stdout]);
@@ -763,8 +763,8 @@ const App = ({
     clearItems();
     clearConsoleMessagesState();
     console.clear();
-    refreshStatic();
-  }, [clearItems, clearConsoleMessagesState, refreshStatic]);
+    refreshConfig();
+  }, [clearItems, clearConsoleMessagesState, refreshConfig]);
 
   const mainControlsRef = useRef<DOMElement>(null);
   const pendingHistoryItemRef = useRef<DOMElement>(null);
@@ -792,20 +792,20 @@ const App = ({
     // debounce so it doesn't fire up too often during resize
     const handler = setTimeout(() => {
       setStaticNeedsRefresh(false);
-      refreshStatic();
+      refreshConfig();
     }, 300);
 
     return () => {
       clearTimeout(handler);
     };
-  }, [terminalWidth, terminalHeight, refreshStatic]);
+  }, [terminalWidth, terminalHeight, refreshConfig]);
 
   useEffect(() => {
     if (streamingState === StreamingState.Idle && staticNeedsRefresh) {
       setStaticNeedsRefresh(false);
-      refreshStatic();
+      refreshConfig();
     }
-  }, [streamingState, refreshStatic, staticNeedsRefresh]);
+  }, [streamingState, refreshConfig, staticNeedsRefresh]);
 
   const filteredConsoleMessages = useMemo(() => {
     if (config.getDebugMode()) {